refactor(article-detail): change functionality to ArticleDetailView
